Se corrigen algunos errores de vista al crear y eliminar registros. Se modifica el template para unir las paginas de la aplicación por medio de un menu tipo pestañas. Se agrega el archivo messages.properties para el manejo de las etiquetas de vista. Se crean los servicios para el login de usuarios con la encriptación de contraseña y el manejo de perfiles. Se implementa el calculo de la edad del alumno en base a su fecha de nacimiento. Se crea enumeración para el manejo del semestre en el catálogo de materias. Se ajusta la ubicación de la carga inicial de catálogos para las vistas y el manejo de la navegación. Se agrega la opción de logout.

+++ TitulacionEJB/ejbModule/mx/upiicsa/titulacion/model/Direccion.java
@@ -16,7 +16,7 @@ public class Direccion implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="IdDireccion")
-	private int idDireccion;
+	private Long idDireccion;
 
 	@Column(name="Calle")
 	private String calle;
@@ -42,11 +42,11 @@ public class Direccion implements Serializable {
 	public Direccion() {
 	}
 
-	public int getIdDireccion() {
+	public Long getIdDireccion() {
 		return this.idDireccion;
 	}
 
-	public void setIdDireccion(int idDireccion) {
+	public void setIdDireccion(Long idDireccion) {
 		this.idDireccion = idDireccion;
 	}
